Extract systemUiController setup into separate composable
