+++ app/src/main/java/com/epmedu/animeal/MainActivity.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.ui.Modifier
 import androidx.core.view.WindowCompat
 import com.epmedu.animeal.foundation.theme.AnimealTheme
+import com.epmedu.animeal.foundation.theme.TransparentSystemUi
 import com.epmedu.animeal.navigation.AnimatedScreenNavHost
 import com.epmedu.animeal.navigation.route.MainRoute
 import com.epmedu.animeal.signup.SignUpScreen
@@ -26,13 +27,15 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             AnimealTheme {
-                AnimatedScreenNavHost(
-                    startDestination = MainRoute.Splash.name,
-                    modifier = Modifier.navigationBarsPadding()
-                ) {
-                    screen(MainRoute.Splash.name) { SplashScreen() }
-                    screen(MainRoute.SignUp.name) { SignUpScreen() }
-                    screen(MainRoute.Tabs.name) { TabsScreen() }
+                TransparentSystemUi {
+                    AnimatedScreenNavHost(
+                        startDestination = MainRoute.Splash.name,
+                        modifier = Modifier.navigationBarsPadding()
+                    ) {
+                        screen(MainRoute.Splash.name) { SplashScreen() }
+                        screen(MainRoute.SignUp.name) { SignUpScreen() }
+                        screen(MainRoute.Tabs.name) { TabsScreen() }
+                    }
                 }
             }
         }
