+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/theme/Theme.kt
@@ -6,10 +6,8 @@ import androidx.compose.material.Surface
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.graphics.Color
 import com.epmedu.animeal.foundation.theme.AnimealColor.Light
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 private val DarkColorPalette = darkColors(
     primary = Light.SeaSerpent,
@@ -37,7 +35,6 @@ fun AnimealTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable () -> Unit
 ) {
-    val systemUiController = rememberSystemUiController()
     val colors = if (darkTheme) {
         DarkColorPalette
     } else {
@@ -49,18 +46,6 @@ fun AnimealTheme(
         typography = Typography,
         shapes = Shapes
     ) {
-        LaunchedEffect(Unit) {
-            systemUiController.run {
-                setNavigationBarColor(
-                    color = Color.Transparent
-                )
-                setSystemBarsColor(
-                    color = Color.Transparent,
-                    darkIcons = darkTheme.not()
-                )
-            }
-        }
-
         Surface(content = content)
     }
 }
